0.145.20180624.1
added: phase 1 of the linux desktop switch changes has arrived
changed: installHachreShell should now work correctly on CentOS, Arch and Ubuntu
added: installHachreShell now also installs nano-syntax-highlighting
changed: installHachreShell can now be called regularily to update nano and zsh syntax highlighting
changed: dy on Arch now uses Yay instead of Pacaur to do AUR stuff
removed: tons of legacy aliases for Arch
refactor: started refactoring efforts in order to leak variables less visibly
changed: pmSetup is now called dySetup (it sets up the dy scripts)
added: several distros now have DistroNames in addition to "DetectedDistro" ids
added: dyl allows dispalying all files belonging to a package

+++ installHachreShell.sh
@@ -1,34 +1,79 @@
 #!/bin/bash
+#
+# 1.0.20180624.1
+#
 
-wget -q https://raw.githubusercontent.com/hachre/aliases/master/source/aliases.sh
-source aliases.sh
-rm aliases.sh
-
-if [ "$dyDetectedDistro" == "CentOS" ]; then
-	yum install -y epel-release || true
-	yum-config-manager --enable epel || true
+if [ $(whoami) != "root" ]; then
+	echo "Error: This setup needs to be run as root."
+	exit 1
 fi
 
-dyx || true
-dyi -y git zsh byobu mosh htop aria2
-rehash 1>/dev/null 2>&1
+which wget 1>/dev/null 2>&1
+if [ "$?" != "0" ]; then
+	echo "Error: We need 'wget' to be installed."
+	exit 1
+fi
 
+# Install hachreAliases
+rm -R /usr/local/hachre/aliases 1>/dev/null 2>&1
+cd /tmp
 wget -q https://raw.githubusercontent.com/hachre/aliases/master/install.sh
 bash install.sh
 rm install.sh
 source /usr/local/hachre/aliases/source/aliases.sh
 
+# CentOS specific prequisities
+if [ "$dyDetectedDistro" == "CentOS" ]; then
+	yum install -y epel-release || true
+	yum-config-manager --enable epel || true
+fi
+
+# Set the noconfirm flag based on the distro in use
+noconfirm=""
+if [ "$dyDetectedDistro" == "CentOS" ]; then
+	noconfirm="-y"
+fi
+if [ "$dyDetectedDistro" == "arch" ]; then
+	noconfirm="--noconfirm"
+fi
+if [ "$dyDetectedDistro" == "ubuntu" ]; then
+	noconfirm="-y"
+fi
+
+# Install the tools we'd like to have
+dyx || true
+dyi "$noconfirm" git zsh byobu mosh htop aria2 curl nano
+rehash 1>/dev/null 2>&1
+
+# Install zsh syntax highlighting
 cd /usr/local/hachre/aliases
 git clone https://github.com/zsh-users/zsh-syntax-highlighting.git
 
-cd /tmp
-wget -q https://raw.githubusercontent.com/hachre/aliases/master/root-skel.tar.gz
-cd /
-tar xvzf /tmp/root-skel.tar.gz && rm /tmp/root-skel.tar.gz
+# Install nano syntax highlighting
+cd /usr/local/hachre/aliases
+git clone https://github.com/scopatz/nanorc.git nano-syntax-highlighting
+cat /etc/nanorc | grep -v "hachre/aliases" > /etc/nanorc.tmp
+mv /etc/nanorc.tmp /etc/nanorc
+echo "include /usr/local/hachre/aliases/nano-syntax-highlighting/*.nanorc" >> /etc/nanorc
+
+# Install the Root Skel
+if [ ! -f "/root/.zshrc_grml" ]; then
+	cd /tmp
+	wget -q https://raw.githubusercontent.com/hachre/aliases/master/root-skel.tar.gz
+	cd /
+	tar xzf /tmp/root-skel.tar.gz && rm /tmp/root-skel.tar.gz
+else
+	echo "Info: You already have a '/root/.zshrc-grml', so installing the Root Skel is skipped..."
+fi
 chown -R root:root /root
 chmod u=rwX,g-rwx,o-rwx /root -R
 
+# Switch to zsh
 chsh -s `which zsh`
 
-echo "Almost done! Please logout and back in to your shell"
-echo "and run 'byobu-enable' followed by 'byobu' to finish."
+# Finished
+echo ""
+echo "Setup complete."
+echo "Please run 'byobu-enable' and then log out and back in again."
+echo ""
+zsh
